// crates/cauce-server/web/src/types/DayCount.ts
// This Source Code Form is subject to the terms of the Mozilla Public
// License, v. 2.0. If a copy of the MPL was not distributed with this
// file, You can obtain one at <https://mozilla.org/MPL/2.0/>.

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Searches per UTC day, for the dashboard chart.
 */
export type DayCount = { day: string, searches: number, cache_hits: number, };